Moving reporting functionality into its own utilities file
