+++ docs/README.md
@@ -4,4 +4,5 @@ This folder contains supplemental documentation to help explain the implementati
 
 |File|Description|
 |-|-|
-|FilePropertyManagement|Documentation on managing default and overriding build properties for specific application files|
+|[FilePropertyManagement](FilePropertyManagement.md)|Documentation on managing default and overriding build properties for specific application files|
+|[Reporting Capabilities](REPORTS.md)|Documentation of zAppBuild reporting capabilities like reporting external impacted file and reporting of concurrent changes|
